saas/ai_api_mgm/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
pydantic>=2
reportlab
pyodbc
requests
